--- CoreData/TrackerCoreDataFactory.swift
@@ -1,14 +1,0 @@
-//
-//  TrackerCoreDataFactory.swift
-//  Tracker
-//
-//  Created by Uliana Lukash on 13.12.2023.
-//
-
-import Foundation
-
-// MARK: - Class
-//
-//final class TrackersCoreDataFactory: NSObject {
-//
-//}
