+++ ViewControllers/CategoriesTableViewCell.swift
@@ -13,7 +13,8 @@ final class CategoriesTableViewCell: UITableViewCell {
     // MARK: - Public Properties
     
     static let identifier = "CategoriesTableViewCell"
-
+    var viewModel: SetCategoryViewModel?
+    
     // MARK: - Private Properties
     
     
@@ -22,8 +23,7 @@ final class CategoriesTableViewCell: UITableViewCell {
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: .default, reuseIdentifier: reuseIdentifier)
         
-        self.detailTextLabel?.font = .systemFont(ofSize: 17, weight: .regular)
-        self.detailTextLabel?.textColor = UIColor(named: "Gray")
+        setup()
     }
     
     required init?(coder: NSCoder) {
@@ -32,7 +32,35 @@ final class CategoriesTableViewCell: UITableViewCell {
     
     // MARK: - Public Methods
     
-
+    func configure(indexPath: IndexPath) {
+        selectionStyle = .none
+        textLabel?.text = viewModel?.categories[indexPath.row].name
+        
+        if viewModel?.categoriesNumber() ==  1 {
+            if indexPath.row == 0 {
+                separatorInset = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: .greatestFiniteMagnitude)
+                layer.cornerRadius = 16
+                layer.maskedCorners = [.layerMinXMaxYCorner, .layerMaxXMaxYCorner]
+            }
+        } else if (viewModel?.categoriesNumber())! > 1 {
+            print("No separator becaue index path = \(indexPath.row)")
+            if indexPath.row == (viewModel?.categoriesNumber())! - 1 {
+                layer.cornerRadius = 16
+                layer.maskedCorners = [.layerMinXMaxYCorner, .layerMaxXMaxYCorner]
+            } else {
+                print("Yes separator becaue index path = \(indexPath.row)")
+                layer.cornerRadius = 0
+            }
+        }
+    }
     // MARK: - Private Methods
 
+    private func setup() {
+        detailTextLabel?.font = .systemFont(ofSize: 17, weight: .regular)
+        detailTextLabel?.textColor = UIColor(named: "Gray")
+        heightAnchor.constraint(equalToConstant: 75).isActive = true
+        backgroundColor = UIColor(named: "Background")
+        
+    }
+    
 }
