+++ ViewControllers/OnboardingViewController.swift
@@ -9,14 +9,14 @@ import Foundation
 import UIKit
 
 final class OnboardingViewController: UIPageViewController {
-    lazy var pages: [UIViewController] = {
+    lazy private var pages: [UIViewController] = {
         let first = OnboardingViewControllerBase(imageName: "background 1", labelText: "Отслеживайте только то, что хотите")
         let second = OnboardingViewControllerBase(imageName: "background 2", labelText: "Даже если это не литры воды и йога")
         
         return [first, second]
     }()
     
-    lazy var pageControl: UIPageControl = {
+    lazy private var pageControl: UIPageControl = {
         let pageControl = UIPageControl()
         pageControl.numberOfPages = pages.count
         pageControl.currentPage = 0
@@ -86,82 +86,3 @@ extension OnboardingViewController: UIPageViewControllerDelegate {
     }
 }
 
-class OnboardingViewControllerBase: UIViewController {
-    
-    let imageName: String?
-    let labelText: String?
-    
-    var imageView: UIImageView = {
-        let imageView = UIImageView(frame: .zero)
-        imageView.contentMode = .scaleToFill
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        return imageView
-    }()
-    
-    let label: UILabel = {
-        let label = UILabel()
-        label.text = "Отслеживайте только то, что хотите"
-        label.font = .systemFont(ofSize: 32, weight: .bold)
-        label.textColor = UIColor(named: "Black")
-        label.textAlignment = .center
-        label.lineBreakMode = .byWordWrapping
-        label.numberOfLines = 0
-        return label
-    }()
-    let button: CustomButton = {
-        let button = CustomButton()
-        button.buttonLabel = "Вот это технологии!"
-        return button
-    }()
-    
-    init(imageName: String?, labelText: String?) {
-        self.imageName = imageName
-        self.labelText = labelText
-        super.init(nibName: nil, bundle: nil)
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        setUp()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    private func setUp() {
-        
-        button.addTarget(self, action: #selector(didTapOnboardingButton), for: .touchUpInside)
-        imageView.image = UIImage(named: imageName!)
-        
-        view.addSubview(button)
-        view.addSubview(label)
-        view.insertSubview(imageView, at: 0)
-        
-        button.translatesAutoresizingMaskIntoConstraints = false
-        label.translatesAutoresizingMaskIntoConstraints = false
-        
-        NSLayoutConstraint.activate([
-            imageView.topAnchor.constraint(equalTo: view.topAnchor),
-            imageView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            imageView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            imageView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            label.leftAnchor.constraint(equalTo: view.leftAnchor, constant: 16),
-            label.rightAnchor.constraint(equalTo: view.rightAnchor, constant: -16),
-            label.centerYAnchor.constraint(equalTo: view.safeAreaLayoutGuide.centerYAnchor),
-            button.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -50),
-            button.leftAnchor.constraint(equalTo: view.leftAnchor, constant: 20),
-            button.rightAnchor.constraint(equalTo: view.rightAnchor, constant: -20)
-        ])
-    }
-    
-    @objc func didTapOnboardingButton() {
-        let viewController = TabBarViewController()
-        let defaults = UserDefaults.standard
-        defaults.set(true, forKey: "SkippedUnboarding")
-        
-        viewController.modalPresentationStyle = .fullScreen
-        present(viewController, animated: true, completion: nil)
-        
-    }
-}
